+++ src/main/java/nextstep/subway/domain/entity/Favorite.java
@@ -1,4 +1,4 @@
-package nextstep.favorite.domain;
+package nextstep.subway.domain.entity;
 
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
