+++ src/main/java/nextstep/subway/dto/FavoriteResponse.java
@@ -1,4 +1,4 @@
-package nextstep.favorite.application.dto;
+package nextstep.subway.dto;
 
 /**
  * TODO: StationResponse를 포함하는 클래스로 만듭니다.
