+++ src/main/java/nextstep/subway/repository/FavoriteRepository.java
@@ -1,5 +1,6 @@
-package nextstep.favorite.domain;
+package nextstep.subway.repository;
 
+import nextstep.subway.domain.entity.Favorite;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FavoriteRepository extends JpaRepository<Favorite, Long> {
