+++ src/main/java/nextstep/subway/service/FavoriteService.java
@@ -1,9 +1,9 @@
-package nextstep.favorite.application;
+package nextstep.subway.service;
 
-import nextstep.favorite.application.dto.FavoriteRequest;
-import nextstep.favorite.application.dto.FavoriteResponse;
-import nextstep.favorite.domain.Favorite;
-import nextstep.favorite.domain.FavoriteRepository;
+import nextstep.subway.dto.FavoriteRequest;
+import nextstep.subway.dto.FavoriteResponse;
+import nextstep.subway.domain.entity.Favorite;
+import nextstep.subway.repository.FavoriteRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
