+++ src/test/java/nextstep/subway/acceptance/FavoriteAcceptanceTest.java
@@ -1,8 +1,10 @@
-package nextstep.favorite.acceptance;
+package nextstep.subway.acceptance;
 
 import nextstep.utils.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
+import org.springframework.boot.test.context.SpringBootTest;
 
 @DisplayName("즐겨찾기 관련 기능")
 public class FavoriteAcceptanceTest extends AcceptanceTest {
+
 }
