+++ src/test/java/nextstep/subway/acceptance/PathAcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance.subway;
+package nextstep.subway.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
