+++ src/test/java/nextstep/subway/acceptance/SectionAcceptanceTest.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance.subway;
+package nextstep.subway.acceptance;
 
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
